Added a basic splash screen.
TODO: change to a more modern theme based splash screen

AndroidManifest.xml
- Changed the splash screen to be the first activity launched

MainActivity.java
- Set the TestFragment to be the default fragment in the MainActivity's
  DrawerLayout when the splash screen is complete

TestFragment.java
- Removed th DB code
- Added a constant that will be used to keep track of data passed to
  this fragment

themes.xml
- Added the splash screen style

colors.xml
- Added colors for the splash screen

strings.xml
- Added additional string resources
