+++ src/main/java/com/nof/vamathcalculator/TestFragment.java
@@ -24,6 +24,12 @@ import android.database.sqlite.SQLiteOpenHelper;
  */
 public class TestFragment extends Fragment {
 
+    /**
+     * Used to get the value to sent to the TestFragment
+     * Key used in the intent passed from LoadingFullScreenActivity.
+     */
+    public static final String TEST_VIEW_DATA = "test_view_data";
+
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -32,7 +38,6 @@ public class TestFragment extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-    private SQLiteOpenHelper db_helper;
 
     public TestFragment() {
         // Required empty public constructor
@@ -61,7 +66,6 @@ public class TestFragment extends Fragment {
     @Override
     public void onAttach(Activity activity) {
        super.onAttach(activity);
-       db_helper = new VAMathDBHelper(activity);
     }
 
     @Override
@@ -86,55 +90,6 @@ public class TestFragment extends Fragment {
         super.onStart();
 
         // Use the getView() method within a fragment to get a reference to its' root element
-
-        new QueryDB().execute(VARates.Basic.Dependent_Status.Alone_No_Depends.getStatus());
-    }
-
-    private class QueryDB extends AsyncTask<String, Void, Double> {
-
-        protected Double doInBackground(String... status) {
-            try {
-                SQLiteDatabase db = db_helper.getReadableDatabase();
-                Cursor cursor = db.query(
-                        VAMathDBHelper.BASIC_TABLE_NAME,
-                        null,
-                        //VAMathDBHelper.DEPEND_STATUS_COLUMN_NAME + " = ?",
-                        //new String[]{status[0]},
-                        null,
-                        null,
-                        null,
-                        null,
-                        null);
-
-                if (cursor.moveToFirst()) {
-                    // skip the first two columns (ID and dependency status)
-                    double query_result = cursor.getDouble(2);
-                    cursor.close();
-                    db.close();
-                    //return (double) cursor.getColumnCount();
-                    return query_result;
-                } else {
-                    cursor.close();
-                    db.close();
-                    return 0.0;
-                }
-
-            } catch (SQLiteException e) {
-                return 0.0;
-            }
-        }
-
-        protected void onPostExecute(Double response) {
-            if (response > 0.0) {
-                View root_view = getView();
-                TextView text_view;
-
-                if (root_view != null) {
-                    text_view = root_view.findViewById(R.id.test_fragment_text);
-                    text_view.setText(Double.toString(response));
-                }
-            }
-        }
 
     }
 }
