+++ src/main/res/values/styles.xml
@@ -1,0 +1,11 @@
+<resources>
+
+    <style name="Widget.Theme.VAMathCalculator.ActionBar.Fullscreen" parent="Widget.AppCompat.ActionBar">
+        <item name="android:background">@color/black_overlay</item>
+    </style>
+
+    <style name="Widget.Theme.VAMathCalculator.ButtonBar.Fullscreen" parent="">
+        <item name="android:background">@color/black_overlay</item>
+        <item name="android:buttonBarStyle">?android:attr/buttonBarStyle</item>
+    </style>
+</resources>
