+++ src/main/res/values/themes.xml
@@ -14,4 +14,17 @@
         <item name="colorAccent">@color/teal_200</item>
         <!-- Customize your theme here. -->
     </style>
+
+    <style name="Theme.VAMathCalculator.Fullscreen" parent="Theme.VAMathCalculator">
+        <item name="android:actionBarStyle">
+            @style/Widget.Theme.VAMathCalculator.ActionBar.Fullscreen
+        </item>
+        <item name="android:windowActionBarOverlay">true</item>
+        <item name="android:windowBackground">@null</item>
+    </style>
+
+    <style name="ThemeOverlay.VAMathCalculator.FullscreenContainer" parent="">
+        <item name="fullscreenBackgroundColor">@color/light_blue_600</item>
+        <item name="fullscreenTextColor">@color/light_blue_A200</item>
+    </style>
 </resources>
